requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
python-dotenv==1.0.0
google-auth-oauthlib==1.1.0
google-auth==2.23.4
google-api-python-client==2.108.0
pydantic==2.5.0